+++ UltimateChartist/UltimateChartist/DataModels/StockSerie.cs
@@ -36,23 +36,23 @@ public class StockSerie
     private DateTime[] dateValues;
     public DateTime[] DateValues => dateValues ??= this.Bars.Select(b => b.Date).ToArray();
 
-    private double[] closeValues;
-    public double[] CloseValues => closeValues ??= this.Bars.Select(b => b.Close).ToArray();
+    private decimal[] closeValues;
+    public decimal[] CloseValues => closeValues ??= this.Bars.Select(b => b.Close).ToArray();
 
-    private double[] highValues;
-    public double[] HighValues => highValues ??= this.Bars.Select(b => b.High).ToArray();
+    private decimal[] highValues;
+    public decimal[] HighValues => highValues ??= this.Bars.Select(b => b.High).ToArray();
 
-    private double[] bodyHighValues;
-    public double[] BodyHighValues => bodyHighValues ??= this.Bars.Select(b => b.BodyHigh).ToArray();
+    private decimal[] bodyHighValues;
+    public decimal[] BodyHighValues => bodyHighValues ??= this.Bars.Select(b => b.BodyHigh).ToArray();
 
-    private double[] lowValues;
-    public double[] LowValues => lowValues ??= this.Bars.Select(b => b.Low).ToArray();
+    private decimal[] lowValues;
+    public decimal[] LowValues => lowValues ??= this.Bars.Select(b => b.Low).ToArray();
 
-    private double[] bodyLowValues;
-    public double[] BodyLowValues => bodyLowValues ??= this.Bars.Select(b => b.BodyLow).ToArray();
+    private decimal[] bodyLowValues;
+    public decimal[] BodyLowValues => bodyLowValues ??= this.Bars.Select(b => b.BodyLow).ToArray();
 
-    private double[] openValues;
-    public double[] OpenValues => openValues ??= this.Bars.Select(b => b.Open).ToArray();
+    private decimal[] openValues;
+    public decimal[] OpenValues => openValues ??= this.Bars.Select(b => b.Open).ToArray();
 
     private long[] volumeValues;
     public long[] VolumeValues => volumeValues ??= this.Bars.Select(b => b.Volume).ToArray();
