+++ UltimateChartist/UltimateChartist/Indicators/IndicatorBase.cs
@@ -12,7 +12,7 @@ public abstract class IndicatorBase : IIndicator
     public string ShortName { get { return this.GetType().Name.Split('_')[1]; } }
     public abstract DisplayType DisplayType { get; }
 
-    public double Max { get; protected set; }
+    public decimal Max { get; protected set; }
 
     /// <summary>
     /// Contains the indicator calculated data
